cleanup to build the forward future always the same way
